% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkedScatter.R
\name{linkedScatter}
\alias{linkedScatter}
\title{linkedScatter}
\usage{
linkedScatter(input, output, session, data, left, right)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{data}{data}

\item{left}{left}

\item{right}{right}
}
\description{
linkedScatter
}
